+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/core/error/FillPercentExceeded.kt
@@ -1,0 +1,3 @@
+package com.ryccoatika.sqatools.fillstorage.core.error
+
+class FillPercentExceeded(val fillPercent: Float) : Exception()