+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/core/model/DummyFile.kt
@@ -1,0 +1,7 @@
+package com.ryccoatika.sqatools.fillstorage.core.model
+
+data class DummyFile(
+  val name: String,
+  val sizeInMB: Double,
+  val path: String,
+)