+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/core/utils/FillStorageTextCreator.kt
@@ -2,6 +2,7 @@ package com.ryccoatika.sqatools.fillstorage.core.utils
 
 import android.content.Context
 import com.ryccoatika.sqatools.fillstorage.R
+import com.ryccoatika.sqatools.fillstorage.core.error.FillPercentExceeded
 import com.ryccoatika.sqatools.fillstorage.core.model.FillStorage
 import com.ryccoatika.sqatools.fillstorage.core.model.Storage
 
@@ -10,88 +11,54 @@ internal class FillStorageTextCreator(
 ) {
   fun errorMessage(t: Throwable): String {
     return when {
+      t is FillPercentExceeded -> context.getString(R.string.fs_error_fill_percent, t.fillPercent)
       else -> t.localizedMessage ?: ""
     }
   }
 
-  fun storageTypeCardTitle(
-    type: Storage.Type,
+  fun storageTitle(
+    storage: Storage,
   ): String {
-    return when (type) {
+    return when (storage.type) {
       is Storage.Type.External -> {
-        "External (${type.name})"
+        context.getString(R.string.fs_title_manage_external, storage.type.name)
       }
 
       Storage.Type.Internal -> {
-        "Internal"
+        context.getString(R.string.fs_title_manage_internal)
       }
 
       Storage.Type.Unknown -> {
-        "Unknown"
+        context.getString(R.string.fs_title_manage_unknown)
       }
     }
   }
 
-  fun storageMetricText(
-    metric: Storage.Metric,
-  ): String {
-    return when (metric) {
-      Storage.Metric.MB -> "MB"
-      Storage.Metric.GB -> "GB"
-    }
-  }
-
-  private fun storageCapacityText(
-    capacity: Float,
-    metric: Storage.Metric,
-    withDecimal: Boolean = true,
-  ): String {
-    val format = if (withDecimal) "%.2f" else "%.0f"
-    val capacityText = format.format(capacity)
-    return "$capacityText ${storageMetricText(metric)}"
-  }
-
-  fun storageFreeSpaceText(
+  fun storageCapacityDesc(
     storage: Storage,
   ): String {
-    return storageCapacityText(
-      capacity = storage.freeSpace,
-      metric = storage.metric,
+    val usedCapacity = context.getString(
+      R.string.fs_text_gb_value,
+      storage.capacityInGB.usedSpace,
     )
+    val totalCapacity = context.getString(
+      R.string.fs_text_gb_value,
+      storage.capacityInGB.totalSpace,
+    )
+
+    return context.getString(R.string.fs_desc_storage_capacity, usedCapacity, totalCapacity)
   }
 
-  fun storageUsedSpaceText(
+  fun storageFreeSpace(
     storage: Storage,
   ): String {
-    return storageCapacityText(
-      capacity = storage.usedSpace,
-      metric = storage.metric,
-    )
+    return context.getString(R.string.fs_text_gb_value, storage.capacityInGB.freeSpace)
   }
 
-  fun storageCapacityDetailText(
+  fun storageUsedSpacePercentage(
     storage: Storage,
   ): String {
-    val usedCapacity = storageCapacityText(
-      capacity = storage.usedSpace,
-      metric = storage.metric,
-    )
-    val totalCapacity = storageCapacityText(
-      capacity = storage.totalSpace,
-      metric = storage.metric,
-    )
-
-    return "$usedCapacity of $totalCapacity"
-  }
-
-  fun storageTypeManageTitle(
-    type: Storage.Type,
-  ): String {
-    return when (type) {
-      is Storage.Type.External -> context.getString(R.string.fs_title_manage_external)
-      Storage.Type.Internal -> context.getString(R.string.fs_title_manage_internal)
-      Storage.Type.Unknown -> context.getString(R.string.fs_title_manage_unknown)
-    }
+    return percentageText(storage.usedSpacePercent)
   }
 
   fun percentageText(value: Float): String {
@@ -99,15 +66,35 @@ internal class FillStorageTextCreator(
   }
 
   private fun percentageText(value: Int): String {
-    return "$value%"
+    return context.getString(R.string.fs_text_percent_value, value)
+  }
+
+  fun storageNonDummyFilesLabel(storage: Storage): String {
+    return context.getString(R.string.fs_text_used_space, storage.capacityInGB.nonDummyFiles)
+  }
+
+  fun storageFreeSpaceLabel(storage: Storage): String {
+    return context.getString(R.string.fs_text_free_space, storage.capacityInGB.freeSpace)
+  }
+
+  fun storageDummyFilesLabel(storage: Storage): String {
+    return context.getString(R.string.fs_text_dummy_files, storage.capacityInGB.dummyFiles)
   }
 
   fun fillStorageButtonText(
     fillStorage: FillStorage,
   ): String {
     return when (fillStorage.type) {
-      FillStorage.Type.MB -> storageCapacityText(fillStorage.value, Storage.Metric.MB, false)
-      FillStorage.Type.GB -> storageCapacityText(fillStorage.value, Storage.Metric.GB, false)
+      FillStorage.Type.MB -> context.getString(
+        R.string.fs_text_mb_value_no_decimal,
+        fillStorage.value,
+      )
+
+      FillStorage.Type.GB -> context.getString(
+        R.string.fs_text_gb_value_no_decimal,
+        fillStorage.value,
+      )
+
       FillStorage.Type.PERCENT -> percentageText(fillStorage.value.toInt())
     }
   }
@@ -115,17 +102,7 @@ internal class FillStorageTextCreator(
   fun fillStorageProgress(
     progress: FillStorage.Progress,
   ): String {
-    val filledSize = storageCapacityText(
-      capacity = progress.mbFilled,
-      metric = Storage.Metric.MB,
-      withDecimal = false,
-    )
-    val toFillSize = storageCapacityText(
-      capacity = progress.mbFill,
-      metric = Storage.Metric.MB,
-      withDecimal = false,
-    )
-    return context.getString(R.string.fs_desc_fill_progress, filledSize, toFillSize)
+    return context.getString(R.string.fs_desc_fill_progress, progress.mbFilled, progress.mbFill)
   }
 
   fun fillStorageProgressButtonText(
@@ -134,6 +111,7 @@ internal class FillStorageTextCreator(
     return when {
       progress.isSuccess ||
         progress.error != null -> context.getString(R.string.fs_button_close)
+
       else -> context.getString(R.string.fs_button_cancel)
     }
   }