+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/ui/FeatureScreens.kt
@@ -1,6 +1,7 @@
 package com.ryccoatika.sqatools.fillstorage.ui
 
 import com.ryccoatika.sqatools.fillstorage.inject.FillStorageScope
+import com.ryccoatika.sqatools.fillstorage.ui.dummyfiles.DummyFiles
 import com.ryccoatika.sqatools.fillstorage.ui.home.Home
 import com.ryccoatika.sqatools.fillstorage.ui.manage.Manage
 import me.tatarka.inject.annotations.Inject
@@ -10,4 +11,5 @@ import me.tatarka.inject.annotations.Inject
 internal class FeatureScreens(
   val home: Home,
   val manage: Manage,
+  val dummyFiles: DummyFiles,
 )