+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/ui/dummyfiles/DummyFilesViewState.kt
@@ -1,0 +1,17 @@
+package com.ryccoatika.sqatools.fillstorage.ui.dummyfiles
+
+import androidx.compose.runtime.Immutable
+import com.ryccoatika.sqatools.fillstorage.core.model.DummyFile
+
+@Immutable
+internal data class DummyFilesViewState(
+  val files: List<DummyFile>,
+  val isLoading: Boolean,
+) {
+  companion object {
+    val Empty = DummyFilesViewState(
+      files = emptyList(),
+      isLoading = false,
+    )
+  }
+}