+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/ui/home/HomeViewModel.kt
@@ -2,12 +2,10 @@ package com.ryccoatika.sqatools.fillstorage.ui.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.ryccoatika.sqatools.fillstorage.core.model.Storage
 import com.ryccoatika.sqatools.fillstorage.core.usecase.ObserveStorages
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import me.tatarka.inject.annotations.Inject
 
@@ -15,27 +13,15 @@ import me.tatarka.inject.annotations.Inject
 internal class HomeViewModel(
   observeStorages: ObserveStorages,
 ) : ViewModel() {
-  private val metric = MutableStateFlow(Storage.Metric.GB)
-
-  val state: StateFlow<HomeViewState> = combine(
-    metric,
-    observeStorages.flow,
-  ) { metric, storages ->
-    HomeViewState(
-      metric = metric,
-      storages = storages.map { it.convert(metric) },
+  val state: StateFlow<HomeViewState> = observeStorages.flow
+    .map(::HomeViewState)
+    .stateIn(
+      scope = viewModelScope,
+      started = SharingStarted.WhileSubscribed(),
+      initialValue = HomeViewState.Empty,
     )
-  }.stateIn(
-    scope = viewModelScope,
-    started = SharingStarted.WhileSubscribed(),
-    initialValue = HomeViewState.Empty,
-  )
 
   init {
     observeStorages(Unit)
-  }
-
-  fun updateMetric(metric: Storage.Metric) {
-    this.metric.value = metric
   }
 }