+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/ui/home/HomeViewState.kt
@@ -5,12 +5,10 @@ import com.ryccoatika.sqatools.fillstorage.core.model.Storage
 
 @Immutable
 internal data class HomeViewState(
-  val metric: Storage.Metric,
   val storages: List<Storage>,
 ) {
   companion object {
     val Empty = HomeViewState(
-      metric = Storage.Metric.GB,
       storages = emptyList(),
     )
   }